Move all splitting methods to SparkDataSplitter
