+++ great_expectations/execution_engine/sparkdf_execution_engine.py
@@ -313,18 +313,10 @@ Please check your config."""
     def _apply_splitting_and_sampling_methods(self, batch_spec, batch_data):
         if batch_spec.get("splitter_method"):
             splitter_method_name: str = batch_spec.get("splitter_method")
-            # TODO: AJB 20220426 move all splitter methods to SparkDataSplitter
-            if splitter_method_name in [
-                "split_on_year",
-                "split_on_year_and_month",
-                "split_on_year_and_month_and_day",
-                "split_on_date_parts",
-            ]:
-                splitter_fn: Callable = self._data_splitter.get_splitter_method(
-                    splitter_method_name
-                )
-            else:
-                splitter_fn: Callable = getattr(self, batch_spec.get("splitter_method"))
+
+            splitter_fn: Callable = self._data_splitter.get_splitter_method(
+                splitter_method_name
+            )
             splitter_kwargs: dict = batch_spec.get("splitter_kwargs") or {}
             batch_data = splitter_fn(batch_data, **splitter_kwargs)
 
@@ -677,108 +669,6 @@ Please check your config."""
     def head(self, n=5):
         """Returns dataframe head. Default is 5"""
         return self.dataframe.limit(n).toPandas()
-
-    @staticmethod
-    def _split_on_whole_table(
-        df,
-    ):
-        return df
-
-    @staticmethod
-    def _split_on_column_value(df, column_name: str, batch_identifiers: dict):
-        return df.filter(F.col(column_name) == batch_identifiers[column_name])
-
-    @staticmethod
-    def _split_on_converted_datetime(
-        df,
-        column_name: str,
-        batch_identifiers: dict,
-        date_format_string: str = "yyyy-MM-dd",
-    ):
-        matching_string = batch_identifiers[column_name]
-        res = (
-            df.withColumn(
-                "date_time_tmp", F.from_unixtime(F.col(column_name), date_format_string)
-            )
-            .filter(F.col("date_time_tmp") == matching_string)
-            .drop("date_time_tmp")
-        )
-        return res
-
-    @staticmethod
-    def _split_on_divided_integer(
-        df, column_name: str, divisor: int, batch_identifiers: dict
-    ):
-        """Divide the values in the named column by `divisor`, and split on that"""
-        matching_divisor = batch_identifiers[column_name]
-        res = (
-            df.withColumn(
-                "div_temp",
-                (F.col(column_name) / divisor).cast(sparktypes.IntegerType()),
-            )
-            .filter(F.col("div_temp") == matching_divisor)
-            .drop("div_temp")
-        )
-        return res
-
-    @staticmethod
-    def _split_on_mod_integer(df, column_name: str, mod: int, batch_identifiers: dict):
-        """Divide the values in the named column by `divisor`, and split on that"""
-        matching_mod_value = batch_identifiers[column_name]
-        res = (
-            df.withColumn(
-                "mod_temp", (F.col(column_name) % mod).cast(sparktypes.IntegerType())
-            )
-            .filter(F.col("mod_temp") == matching_mod_value)
-            .drop("mod_temp")
-        )
-        return res
-
-    @staticmethod
-    def _split_on_multi_column_values(df, column_names: list, batch_identifiers: dict):
-        """Split on the joint values in the named columns"""
-        for column_name in column_names:
-            value = batch_identifiers.get(column_name)
-            if not value:
-                raise ValueError(
-                    f"In order for SparkDFExecutionEngine to `_split_on_multi_column_values`, "
-                    f"all values in  column_names must also exist in batch_identifiers. "
-                    f"{column_name} was not found in batch_identifiers."
-                )
-            df = df.filter(F.col(column_name) == value)
-        return df
-
-    @staticmethod
-    def _split_on_hashed_column(
-        df,
-        column_name: str,
-        hash_digits: int,
-        batch_identifiers: dict,
-        hash_function_name: str = "sha256",
-    ):
-        """Split on the hashed value of the named column"""
-        try:
-            getattr(hashlib, hash_function_name)
-        except (TypeError, AttributeError):
-            raise (
-                ge_exceptions.ExecutionEngineError(
-                    f"""The splitting method used with SparkDFExecutionEngine has a reference to an invalid hash_function_name.
-                    Reference to {hash_function_name} cannot be found."""
-                )
-            )
-
-        def _encrypt_value(to_encode):
-            hash_func = getattr(hashlib, hash_function_name)
-            hashed_value = hash_func(to_encode.encode()).hexdigest()[-1 * hash_digits :]
-            return hashed_value
-
-        encrypt_udf = F.udf(_encrypt_value, sparktypes.StringType())
-        res = (
-            df.withColumn("encrypted_value", encrypt_udf(column_name))
-            .filter(F.col("encrypted_value") == batch_identifiers["hash_value"])
-            .drop("encrypted_value")
-        )
-        return res
 
     ### Sampling methods ###
     @staticmethod
